feat(utils): move build domain method to utils
