+++ client/options.go
@@ -3,21 +3,15 @@ package client
 import (
 	"flag"
 	"fmt"
-	"net/url"
 	"os"
 
 	"github.com/omarahm3/squirrel/utils"
 	"go.uber.org/zap/zapcore"
 )
 
-type Domain struct {
-	Public    string
-	Websocket string
-}
-
 type ClientOptions struct {
 	Env      string
-	Domain   *Domain
+	Domain   *utils.Domain
 	LogLevel zapcore.Level
 }
 
@@ -37,29 +31,6 @@ func fprintf(format string, a ...interface{}) {
 	fmt.Fprintf(os.Stderr, format, a...)
 }
 
-func buildDomain(domain string, env string) *Domain {
-	var s string
-
-	if env == "prod" {
-		s = "s"
-	}
-
-	public := url.URL{
-		Scheme: fmt.Sprintf("http%s", s),
-		Host:   domain,
-	}
-
-	websocket := url.URL{
-		Scheme: fmt.Sprintf("ws%s", s),
-		Host:   domain,
-	}
-
-	return &Domain{
-		Public:    public.String(),
-		Websocket: websocket.String(),
-	}
-}
-
 func InitOptions() *ClientOptions {
 	flag.Usage = func() {
 		fprintf("Usage of %s:\n", os.Args[0])
@@ -75,7 +46,7 @@ func InitOptions() *ClientOptions {
 
 	return &ClientOptions{
 		Env:      env,
-		Domain:   buildDomain(domain, env),
+		Domain:   utils.BuildDomain(domain, env),
 		LogLevel: utils.GetLogLevelFromString(loglevel),
 	}
 }
